+++ source/test.js
@@ -1,1 +1,15 @@
-console.log('no tests yet!')
+const progress = require('../').create().step('the task you are currently performing')
+// use an array of steps that execute one second after each other
+// as if we do them all instantly
+// you won't see the progress bar as it will be instant
+;[
+	() => progress.setTotal(5),
+	() => progress.setTick(1),
+	() => progress.setTick(2),
+	() => progress.setTick(3),
+	() => progress.addTick(),
+	() => progress.addTick(),
+	() => progress.finish()  // remove and destroy the progress bar
+].forEach(function (step, index) {
+	setTimeout(step, index * 1000)
+})