+++ HISTORY.md
@@ -1,5 +1,9 @@
 ## History
 
+## v1.1.1 2016 May 4
+- Fixed `Range Error: Invalid array length` that could occur under special circumstances
+  - Thanks to [Steve McArthur](https://github.com/SteveMcArthur) for [issue #4](https://github.com/bevry/progressbar/issues/4)
+
 ## v1.1.0 2016 May 2
 - Converted from CoffeeScript to JavaScript